Refactor GRPCDestination func
